fix(quality): decompose the Case/Consultation service cluster (6 phpmd findings)
Clears all six remaining ExcessiveClassComplexity findings in the case and
consultation services by extracting cohesive collaborators — no suppressions,
no baselines, no threshold changes.

  CaseEmailService      80 -> 31
  CaseRelationService   86 -> 39
  CaseSharingService   109 -> 25
  CaseTransferService   63 -> 40
  ConsultationService   77 -> 40
  CaseModelEngine      117 -> 30

15 new collaborators, each well under the threshold (largest: 41). All method
bodies moved verbatim; the only behavioural-surface change is that
CaseEmailRepository::recordSentEmail() receives the already-resolved envelope
from-address as a parameter instead of re-reading the same app-config key.

Removes the class-level @SuppressWarnings(PHPMD.ExcessiveClassComplexity) from
CaseModelEngine, which was ineffective anyway (it sat on the file docblock, not
the class docblock) and is now unnecessary.

phpmd 25 -> 19 (exactly -6, no new finding anywhere in lib/); phpunit
1684/0 failures unchanged; phpcs, psalm, phpstan all exit 0.
