+++ lib/Service/CaseRelationService.php
@@ -37,8 +37,9 @@ declare(strict_types=1);
 
 namespace OCA\Procest\Service;
 
-use OCA\Procest\Service\Support\SearchesObjects;
-use Psr\Log\LoggerInterface;
+use OCA\Procest\Service\Relation\CaseHierarchyOverlapGuard;
+use OCA\Procest\Service\Relation\CaseRelationCodec;
+use OCA\Procest\Service\Relation\CaseRelationStore;
 
 /**
  * Service for typed peer relations between cases.
@@ -47,8 +48,6 @@ use Psr\Log\LoggerInterface;
  */
 class CaseRelationService
 {
-
-    use SearchesObjects;
 
     /**
      * Allowed ZRC relation types (`aardRelatie`).
@@ -60,12 +59,14 @@ class CaseRelationService
     /**
      * Constructor.
      *
-     * @param SettingsService $settingsService Shared OR/settings resolver.
-     * @param LoggerInterface $logger          Logger.
+     * @param CaseRelationStore         $store          OpenRegister reads/writes for case objects.
+     * @param CaseRelationCodec         $codec          Relation-list encoding and pair operations.
+     * @param CaseHierarchyOverlapGuard $hierarchyGuard Hoofdzaak/deelzaak overlap detection.
      */
     public function __construct(
-        private readonly SettingsService $settingsService,
-        private readonly LoggerInterface $logger,
+        private readonly CaseRelationStore $store,
+        private readonly CaseRelationCodec $codec,
+        private readonly CaseHierarchyOverlapGuard $hierarchyGuard,
     ) {
     }//end __construct()
 
@@ -84,12 +85,12 @@ class CaseRelationService
      */
     public function listRelations(string $caseId): array
     {
-        $case = $this->fetchCase(caseUuid: $caseId);
+        $case = $this->store->fetchCase(caseUuid: $caseId);
         if ($case === null) {
             return [];
         }
 
-        return $this->decodeRelations(case: $case);
+        return $this->codec->decode(case: $case);
     }//end listRelations()
 
     /**
@@ -129,15 +130,15 @@ class CaseRelationService
         }
 
         // OR-RBAC read access to BOTH cases (fail closed on either miss).
-        $origin = $this->fetchCase(caseUuid: $caseId);
-        $target = $this->fetchCase(caseUuid: $targetId);
+        $origin = $this->store->fetchCase(caseUuid: $caseId);
+        $target = $this->store->fetchCase(caseUuid: $targetId);
         if ($origin === null || $target === null) {
             return ['ok' => false, 'reason' => 'access_denied'];
         }
 
         // Hierarchy-overlap guard — the parent/sub-case link already expresses
         // the relation, so refuse to also peer-link the same pair.
-        if ($this->isHierarchyLinked(caseA: $origin, caseB: $target) === true) {
+        if ($this->hierarchyGuard->areLinked(caseA: $origin, caseB: $target) === true) {
             return [
                 'ok'     => false,
                 'reason' => 'hierarchy_overlap',
@@ -145,17 +146,17 @@ class CaseRelationService
             ];
         }
 
-        $originRelations = $this->decodeRelations(case: $origin);
-        if ($this->hasPair(relations: $originRelations, caseId: $targetId, aardRelatie: $aardRelatie) === true) {
+        $originRelations = $this->codec->decode(case: $origin);
+        if ($this->codec->hasPair(relations: $originRelations, caseId: $targetId, aardRelatie: $aardRelatie) === true) {
             return ['ok' => false, 'reason' => 'duplicate'];
         }
 
-        $originRelations[] = $this->buildRelationEntry(
+        $originRelations[] = $this->codec->buildEntry(
             caseId: $targetId,
             aardRelatie: $aardRelatie,
             toelichting: $toelichting
         );
-        $this->persistRelations(case: $origin, relations: $originRelations);
+        $this->store->persistRelations(case: $origin, relations: $originRelations);
 
         // Symmetric counterpart — same type names the link, the UI renders
         // direction-aware labels.
@@ -199,25 +200,6 @@ class CaseRelationService
     }//end rejectInvalidRelationInput()
 
     /**
-     * Build a single relation entry, carrying the optional clarification.
-     *
-     * @param string      $caseId      Referenced case UUID.
-     * @param string      $aardRelatie Relation type.
-     * @param string|null $toelichting Optional free-text clarification.
-     *
-     * @return array<string, string>
-     */
-    private function buildRelationEntry(string $caseId, string $aardRelatie, ?string $toelichting): array
-    {
-        $entry = ['caseId' => $caseId, 'aardRelatie' => $aardRelatie];
-        if ($toelichting !== null && $toelichting !== '') {
-            $entry['toelichting'] = $toelichting;
-        }
-
-        return $entry;
-    }//end buildRelationEntry()
-
-    /**
      * Persist the symmetric counterpart entry on the target case, unless it is
      * already present.
      *
@@ -234,14 +216,14 @@ class CaseRelationService
         string $aardRelatie,
         ?string $toelichting
     ): void {
-        $targetRelations = $this->decodeRelations(case: $target);
-        if ($this->hasPair(relations: $targetRelations, caseId: $caseId, aardRelatie: $aardRelatie) === false) {
-            $targetRelations[] = $this->buildRelationEntry(
+        $targetRelations = $this->codec->decode(case: $target);
+        if ($this->codec->hasPair(relations: $targetRelations, caseId: $caseId, aardRelatie: $aardRelatie) === false) {
+            $targetRelations[] = $this->codec->buildEntry(
                 caseId: $caseId,
                 aardRelatie: $aardRelatie,
                 toelichting: $toelichting
             );
-            $this->persistRelations(case: $target, relations: $targetRelations);
+            $this->store->persistRelations(case: $target, relations: $targetRelations);
         }
     }//end addInverseRelation()
 
@@ -262,25 +244,25 @@ class CaseRelationService
             return ['ok' => false, 'reason' => 'missing_case_id'];
         }
 
-        $origin = $this->fetchCase(caseUuid: $caseId);
-        $target = $this->fetchCase(caseUuid: $targetId);
+        $origin = $this->store->fetchCase(caseUuid: $caseId);
+        $target = $this->store->fetchCase(caseUuid: $targetId);
         if ($origin === null || $target === null) {
             return ['ok' => false, 'reason' => 'access_denied'];
         }
 
-        $originRelations = $this->removePair(
-            relations: $this->decodeRelations(case: $origin),
+        $originRelations = $this->codec->removePair(
+            relations: $this->codec->decode(case: $origin),
             caseId: $targetId,
             aardRelatie: $aardRelatie
         );
-        $this->persistRelations(case: $origin, relations: $originRelations);
+        $this->store->persistRelations(case: $origin, relations: $originRelations);
 
-        $targetRelations = $this->removePair(
-            relations: $this->decodeRelations(case: $target),
+        $targetRelations = $this->codec->removePair(
+            relations: $this->codec->decode(case: $target),
             caseId: $caseId,
             aardRelatie: $aardRelatie
         );
-        $this->persistRelations(case: $target, relations: $targetRelations);
+        $this->store->persistRelations(case: $target, relations: $targetRelations);
 
         return ['ok' => true];
     }//end removeRelation()
@@ -304,12 +286,12 @@ class CaseRelationService
             return 0;
         }
 
-        $deleted = $this->fetchCase(caseUuid: $caseId);
+        $deleted = $this->store->fetchCase(caseUuid: $caseId);
         // Even when the case is already gone we still scan counterparts: the
         // relation entries on OTHER cases are what must be cleaned up.
         $counterpartIds = [];
         if ($deleted !== null) {
-            foreach ($this->decodeRelations(case: $deleted) as $relation) {
+            foreach ($this->codec->decode(case: $deleted) as $relation) {
                 $ref = (string) ($relation['caseId'] ?? '');
                 if ($ref !== '' && in_array($ref, $counterpartIds, true) === false) {
                     $counterpartIds[] = $ref;
@@ -319,21 +301,16 @@ class CaseRelationService
 
         $updated = 0;
         foreach ($counterpartIds as $counterpartId) {
-            $counterpart = $this->fetchCase(caseUuid: $counterpartId);
+            $counterpart = $this->store->fetchCase(caseUuid: $counterpartId);
             if ($counterpart === null) {
                 continue;
             }
 
-            $relations = $this->decodeRelations(case: $counterpart);
-            $stripped  = array_values(
-                array_filter(
-                    $relations,
-                    static fn (array $relation): bool => (string) ($relation['caseId'] ?? '') !== $caseId
-                )
-            );
+            $relations = $this->codec->decode(case: $counterpart);
+            $stripped  = $this->codec->removeAllForCase(relations: $relations, caseId: $caseId);
 
             if (count($stripped) !== count($relations)) {
-                $this->persistRelations(case: $counterpart, relations: $stripped);
+                $this->store->persistRelations(case: $counterpart, relations: $stripped);
                 $updated++;
             }
         }//end foreach
@@ -361,12 +338,12 @@ class CaseRelationService
             return;
         }
 
-        $case = $this->fetchCase(caseUuid: $caseId);
+        $case = $this->store->fetchCase(caseUuid: $caseId);
         if ($case === null) {
             return;
         }
 
-        foreach ($this->decodeRelations(case: $case) as $relation) {
+        foreach ($this->codec->decode(case: $case) as $relation) {
             $targetId    = (string) ($relation['caseId'] ?? '');
             $aardRelatie = (string) ($relation['aardRelatie'] ?? '');
             if ($targetId === '' || $targetId === $caseId
@@ -375,296 +352,16 @@ class CaseRelationService
                 continue;
             }
 
-            $target = $this->fetchCase(caseUuid: $targetId);
+            $target = $this->store->fetchCase(caseUuid: $targetId);
             if ($target === null) {
                 continue;
             }
 
-            $targetRelations = $this->decodeRelations(case: $target);
-            if ($this->hasPair(relations: $targetRelations, caseId: $caseId, aardRelatie: $aardRelatie) === false) {
+            $targetRelations = $this->codec->decode(case: $target);
+            if ($this->codec->hasPair(relations: $targetRelations, caseId: $caseId, aardRelatie: $aardRelatie) === false) {
                 $targetRelations[] = ['caseId' => $caseId, 'aardRelatie' => $aardRelatie];
-                $this->persistRelations(case: $target, relations: $targetRelations);
+                $this->store->persistRelations(case: $target, relations: $targetRelations);
             }
         }//end foreach
     }//end normalise()
-
-    /**
-     * Determine whether two cases are already linked through the deelzaak
-     * (parent/child) hierarchy in either direction.
-     *
-     * @param array<string, mixed> $caseA First case object.
-     * @param array<string, mixed> $caseB Second case object.
-     *
-     * @return bool
-     */
-    private function isHierarchyLinked(array $caseA, array $caseB): bool
-    {
-        $idA = (string) ($caseA['id'] ?? ($caseA['@self']['id'] ?? ''));
-        $idB = (string) ($caseB['id'] ?? ($caseB['@self']['id'] ?? ''));
-
-        $parentA = $this->parentRef(case: $caseA);
-        $parentB = $this->parentRef(case: $caseB);
-
-        if ($idB !== '' && $parentA === $idB) {
-            return true;
-        }
-
-        if ($idA !== '' && $parentB === $idA) {
-            return true;
-        }
-
-        return false;
-    }//end isHierarchyLinked()
-
-    /**
-     * Read the `parentCase` reference UUID out of a case array (scalar or
-     * expanded-object shape).
-     *
-     * @param array<string, mixed> $case Case object.
-     *
-     * @return string Parent UUID or '' when absent.
-     */
-    private function parentRef(array $case): string
-    {
-        $parent = ($case['parentCase'] ?? null);
-        if (is_string($parent) === true) {
-            return $parent;
-        }
-
-        if (is_array($parent) === true) {
-            $ref = ($parent['id'] ?? ($parent['uuid'] ?? ''));
-            if (is_string($ref) === true) {
-                return $ref;
-            }
-
-            return '';
-        }
-
-        return '';
-    }//end parentRef()
-
-    /**
-     * Decode the JSON-encoded `relatedCases` field into a list of relation
-     * entries, tolerating an already-array shape.
-     *
-     * @param array<string, mixed> $case Case object.
-     *
-     * @return array<int, array<string, mixed>>
-     */
-    private function decodeRelations(array $case): array
-    {
-        $entries = [];
-        foreach ($this->rawRelationList(case: $case) as $item) {
-            if (is_array($item) === false) {
-                continue;
-            }
-
-            $entry = $this->decodeRelationEntry(item: $item);
-            if ($entry === null) {
-                continue;
-            }
-
-            $entries[] = $entry;
-        }//end foreach
-
-        return $entries;
-    }//end decodeRelations()
-
-    /**
-     * Read the raw `relatedCases` payload as a list, accepting either the
-     * JSON-encoded string shape or an already-decoded array.
-     *
-     * @param array<string, mixed> $case Case object.
-     *
-     * @return array<mixed> The raw relation list, or [] when unusable.
-     */
-    private function rawRelationList(array $case): array
-    {
-        $raw  = ($case['relatedCases'] ?? null);
-        $list = [];
-        if (is_array($raw) === true) {
-            $list = $raw;
-        }
-
-        if (is_string($raw) === true && $raw !== '') {
-            $decoded = json_decode($raw, true);
-            if (is_array($decoded) === true) {
-                $list = $decoded;
-            }
-        }
-
-        return $list;
-    }//end rawRelationList()
-
-    /**
-     * Normalise one raw relation item into a relation entry.
-     *
-     * @param array<string, mixed> $item Raw relation item.
-     *
-     * @return array<string, string>|null The entry, or null when it names no case.
-     */
-    private function decodeRelationEntry(array $item): ?array
-    {
-        $targetId = (string) ($item['caseId'] ?? '');
-        if ($targetId === '') {
-            return null;
-        }
-
-        $entry = [
-            'caseId'      => $targetId,
-            'aardRelatie' => (string) ($item['aardRelatie'] ?? ''),
-        ];
-        if (isset($item['toelichting']) === true && (string) $item['toelichting'] !== '') {
-            $entry['toelichting'] = (string) $item['toelichting'];
-        }
-
-        return $entry;
-    }//end decodeRelationEntry()
-
-    /**
-     * Whether a `{caseId, aardRelatie}` pair already exists in a relation list.
-     *
-     * @param array<int, array<string, mixed>> $relations   Relation entries.
-     * @param string                           $caseId      Target case UUID.
-     * @param string                           $aardRelatie Relation type.
-     *
-     * @return bool
-     */
-    private function hasPair(array $relations, string $caseId, string $aardRelatie): bool
-    {
-        foreach ($relations as $relation) {
-            if ((string) ($relation['caseId'] ?? '') === $caseId
-                && (string) ($relation['aardRelatie'] ?? '') === $aardRelatie
-            ) {
-                return true;
-            }
-        }
-
-        return false;
-    }//end hasPair()
-
-    /**
-     * Return a copy of the relation list with the given pair removed.
-     *
-     * @param array<int, array<string, mixed>> $relations   Relation entries.
-     * @param string                           $caseId      Target case UUID.
-     * @param string                           $aardRelatie Relation type.
-     *
-     * @return array<int, array<string, mixed>>
-     */
-    private function removePair(array $relations, string $caseId, string $aardRelatie): array
-    {
-        return array_values(
-            array_filter(
-                $relations,
-                static fn (array $relation): bool => (
-                    (string) ($relation['caseId'] ?? '') !== $caseId
-                    || (string) ($relation['aardRelatie'] ?? '') !== $aardRelatie
-                )
-            )
-        );
-    }//end removePair()
-
-    /**
-     * Fetch a single case object by UUID through the session's ObjectService.
-     *
-     * Resolving via OpenRegister applies its per-object RBAC for the current
-     * user, so an unreadable case resolves to null — this is the access guard.
-     *
-     * @param string $caseUuid Case UUID.
-     *
-     * @return array<string, mixed>|null
-     */
-    private function fetchCase(string $caseUuid): ?array
-    {
-        if ($caseUuid === '') {
-            return null;
-        }
-
-        $objectService = $this->settingsService->getObjectService();
-        if ($objectService === null) {
-            return null;
-        }
-
-        $register = $this->settingsService->getConfigValue('register');
-        $schema   = $this->settingsService->getConfigValue('case_schema');
-        if ($register === '' || $schema === '') {
-            return null;
-        }
-
-        try {
-            $obj = $objectService->find($caseUuid, register: $register, schema: $schema);
-        } catch (\Throwable $e) {
-            $this->logger->debug(
-                'CaseRelationService: case lookup failed',
-                ['uuid' => $caseUuid, 'error' => $e->getMessage()]
-            );
-            return null;
-        }
-
-        return $this->normalizeCaseObject(object: $obj);
-    }//end fetchCase()
-
-    /**
-     * Normalise an OpenRegister lookup result to a plain case array.
-     *
-     * @param mixed $object The value returned by the ObjectService.
-     *
-     * @return array<string, mixed>|null The case as an array, or null when unusable.
-     */
-    private function normalizeCaseObject(mixed $object): ?array
-    {
-        if ($object === null) {
-            return null;
-        }
-
-        if (is_object($object) === true && method_exists($object, 'jsonSerialize') === true) {
-            $object = $object->jsonSerialize();
-        }
-
-        if (is_array($object) === true) {
-            return $object;
-        }
-
-        return null;
-    }//end normalizeCaseObject()
-
-    /**
-     * Persist a relation list back onto a case, JSON-encoding the field
-     * (the `relatedCases` field is a JSON-encoded string).
-     *
-     * @param array<string, mixed>             $case      Case object to update.
-     * @param array<int, array<string, mixed>> $relations Relation entries.
-     *
-     * @return void
-     */
-    private function persistRelations(array $case, array $relations): void
-    {
-        $objectService = $this->settingsService->getObjectService();
-        if ($objectService === null) {
-            return;
-        }
-
-        $register = $this->settingsService->getConfigValue('register');
-        $schema   = $this->settingsService->getConfigValue('case_schema');
-        if ($register === '' || $schema === '') {
-            return;
-        }
-
-        $payload = $case;
-        $payload['relatedCases'] = json_encode(array_values($relations));
-
-        try {
-            $objectService->saveObject(
-                object: $payload,
-                register: $register,
-                schema: $schema,
-            );
-        } catch (\Throwable $e) {
-            $this->logger->warning(
-                'CaseRelationService: failed to persist relatedCases',
-                ['error' => $e->getMessage()]
-            );
-        }
-    }//end persistRelations()
 }//end class
