+++ tests/Unit/Service/CaseRelationServiceTest.php
@@ -27,6 +27,9 @@ declare(strict_types=1);
 namespace OCA\Procest\Tests\Unit\Service;
 
 use OCA\Procest\Service\CaseRelationService;
+use OCA\Procest\Service\Relation\CaseHierarchyOverlapGuard;
+use OCA\Procest\Service\Relation\CaseRelationCodec;
+use OCA\Procest\Service\Relation\CaseRelationStore;
 use OCA\Procest\Service\SettingsService;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
@@ -142,9 +145,14 @@ class CaseRelationServiceTest extends TestCase
         $objectService = $this->makeObjectService($store);
         $this->settingsService->method('getObjectService')->willReturn($objectService);
 
+        // The store, codec and hierarchy guard are real collaborators, not
+        // mocks: every assertion below is about behaviour they inherited
+        // verbatim from CaseRelationService, and the store is still driven
+        // entirely by the mocked SettingsService above.
         return new CaseRelationService(
-            settingsService: $this->settingsService,
-            logger: $this->logger,
+            store: new CaseRelationStore($this->settingsService, $this->logger),
+            codec: new CaseRelationCodec(),
+            hierarchyGuard: new CaseHierarchyOverlapGuard(),
         );
     }//end makeService()
